[LI-HOTFIX] LIKAFKA-18349 - Reuse the same UpdateMetadataRequest object and struct to reduce controller memory usage TICKET = KAFKA-7186 LI_DESCRIPTION =
During controller failover and broker changes, it sends out UpdateMetadataRequest to all brokers in the cluster containing the states for all partitions and live brokers. The current implementation will instantiate the UpdateMetadataRequest object and its serialized form (Struct) for # of brokers times, which causes OOM if the memory exceeds the configure JVM heap size before GC kicks in. We have seen this issue in the production environment for multiple times.
For example, if we have 100 brokers in the cluster and each broker is the leader of 2k partitions, the extra memory usage introduced by controller trying to send out UpdateMetadataRequest is around:
```
<memory used by UpdateMetadataRequest Structs> * <# of brokers> * <total # of leader parittions> = 250B * 100 * 200k = 5GB
```
This patch avoids the recreation of UpdateMetadataReuqest and struct objects if the same builder object is used to build the request. This can significantly reduce memory footprint in Controller based on the above calculation.

RB=1364716
BUG=LIKAFKA-18349
G=Kafka-Code-Reviews
R=luwang,okaraman,mgharat
A=mgharat

EXIT_CRITERIA = TICKET [KAFKA-7186]