[LI-HOTFIX] Implementation of memory pool based on weak references. (#49) LI_DESCRIPTION = * Implementation of memory pool based on weak references.
This will act as a NO-OP memory pool (current behavior) in all the clients and kafka broker except
when there is an opportunity to reuse buffer before it could have been garbage collected.
EXIT_CRITERIA =

Note: It is possible that if we allocate buffers in a strictly increasing order, we may accumulate certain keys over time with their weak references set to null. The idea is that allocation distribution is usually uniform and eventually an allocation for a smaller byte buffer would kick out these keys from the cache.