[LI-HOTFIX] Reject metadata updates when cluster id changes (#92)
TICKET =
LI_DESCRIPTION =
When remove all brokers from one cluster and later add the broker(s) to another cluster, we see the consumers start consuming from the new cluster. This is because we don't do cluster id validation upon metadata response, and causes unexpected cross-consumption for some apps. The fix is to throw an exception and close the underlying networkclient when client receives metadata from a different cluster.

EXIT_CRITERIA = Never since this is a linkedin internal change
Co-authored-by: Ke Hu <kehu@kehu-mn2.linkedin.biz>